Adding currency to salaryhistory and refactor valueobject, migrations

+++ src/SalaryHistory/Infrastructure/Migrations/2024_10_06_073827_create_salary_histories_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('user_id')->index();
             $table->date('on_date')->index();
-            $table->integer('salary')->default(0);
+            $table->float('salary')->default(0);
             $table->string('note')->nullable();
             $table->timestamps();
         });
